+++ ctx_test.go
@@ -20,7 +20,7 @@ func TestSetWorkerIdx(t *testing.T) {
 	})
 
 	t.Run("idx is set", func(t *testing.T) {
-		ctx := SetWorkerIdx(context.Background(), 99)
+		ctx := setWorkerIdx(context.Background(), 99)
 		idx := GetWorkerIdx(ctx)
 		assert.Equal(t, 99, idx)
 	})